+++ bespokeasm/assembler/line_object/__init__.py
@@ -1,7 +1,18 @@
+import re
+
 from bespokeasm.assembler.line_identifier import LineIdentifier
 from bespokeasm.assembler.label_scope import LabelScope
 
 INSTRUCTION_EXPRESSION_PATTERN = '(?:(?:\%|\$|b|0x|\.)?[\w]+|[\+\-\*\/\&\|\^\(\)]|(?:[ \t]*)(?!(?:[ \t]*\;|[ \t]*\v)))+'
+
+PATTERN_ALLOWED_LABELS = re.compile(
+        r'^(?!__|\.\.)(?:(?:\.|_|[a-zA-Z])[a-zA-Z0-9_]*)$',
+        flags=re.IGNORECASE|re.MULTILINE
+    )
+
+def is_valid_label(s: str):
+    res = re.search(PATTERN_ALLOWED_LABELS, s)
+    return (res is not None)
 
 class LineObject:
     def __init__(self, line_id: LineIdentifier, instruction: str, comment: str):